cloudchamber: Start `cloudchamber apply`, the new command to deploy container app changes
This command is able to take the [[container-app]] configurations, and
deploy them to Cloudchamber.

To render the differences, we are introducing a new dependency with
"diff". This was already included in the pnpm lock, however we could
consider not rolling out a new dependency into wrangler unless
absolutely necessary.

The command is designed to be CI friendly. In the tests there is some
example command renders from different kind of configurations.